P2.6 Generate TypeScript Types
## Summary

### Changes Made

1. **Fixed typeshare compatibility issues in Rust contracts:**
   - Removed `#[typeshare]` attribute from `ContentBlock` and `UnifiedAgentEvent` enums (internally tagged enums with `#[serde(tag = "type")]` are not supported by typeshare)
   - Changed `i64`/`u64` types to `i32` in:
     - `AgentStats`: `input_tokens`, `output_tokens`, `total_tokens`, `duration_ms`
     - `EventEnvelope<T>`: `sequence`
     - `AgentEventRecord`: `sequence`

2. **Added manual TypeScript types in `packages/generated/types-manual.ts`:**
   - `ContentBlock` union type with variants: `ContentBlockText`, `ContentBlockToolUse`, `ContentBlockToolResult`, `ContentBlockThinking`
   - `UnifiedAgentEvent` union type with 7 variants: `Init`, `Message`, `ToolUse`, `ToolResult`, `Complete`, `Error`, `Permission`

3. **Generated all TypeScript types:**
   - `pnpm generate:all` runs successfully
   - All 132 schemas generated
   - All 81 endpoints with 16 query files generated

### Files Modified
- `crates/openflow-contracts/src/events/agent_event.rs` - Removed typeshare, fixed i64→i32
- `crates/openflow-contracts/src/events/envelope.rs` - Fixed u64→i32, i64→i32
- `packages/generated/types-manual.ts` - Added ContentBlock and UnifiedAgentEvent types

### Notes
- The frontend has ~90 type errors related to `TaskStatus` enum changes from step P2.3 (old: Todo/Inprogress/Inreview/Done → new: Pending/Running/Paused/Completed/Failed/Cancelled). These are expected and will be fixed in Phase 7 (Frontend) and Phase 8 (Cleanup) of the implementation plan.

+++ crates/openflow-contracts/src/events/agent_event.rs
@@ -79,7 +79,9 @@ impl std::str::FromStr for AgentMessageRole {
 ///
 /// Messages can contain multiple content blocks of different types.
 /// This allows rich content like text mixed with tool results.
-#[typeshare]
+///
+/// Note: This uses internally-tagged serde which typeshare doesn't support.
+/// The TypeScript type is manually defined in types-manual.ts.
 #[derive(Debug, Clone, Serialize, Deserialize, PartialEq)]
 #[serde(tag = "type", rename_all = "snake_case")]
 pub enum ContentBlock {
@@ -279,15 +281,15 @@ impl CompletionStatus {
 #[serde(rename_all = "camelCase")]
 pub struct AgentStats {
     /// Number of input tokens used
-    pub input_tokens: Option<i64>,
+    pub input_tokens: Option<i32>,
     /// Number of output tokens used
-    pub output_tokens: Option<i64>,
+    pub output_tokens: Option<i32>,
     /// Total tokens used
-    pub total_tokens: Option<i64>,
+    pub total_tokens: Option<i32>,
     /// Number of tool calls made
     pub tool_calls: Option<i32>,
     /// Duration in milliseconds
-    pub duration_ms: Option<i64>,
+    pub duration_ms: Option<i32>,
     /// Model used for this session
     pub model: Option<String>,
 }
@@ -299,7 +301,7 @@ impl AgentStats {
     }
 
     /// Create stats with token counts
-    pub fn with_tokens(input: i64, output: i64) -> Self {
+    pub fn with_tokens(input: i32, output: i32) -> Self {
         Self {
             input_tokens: Some(input),
             output_tokens: Some(output),
@@ -385,7 +387,9 @@ impl PermissionRequest {
 /// ```json
 /// {"type": "message", "role": "assistant", "content": [...]}
 /// ```
-#[typeshare]
+///
+/// Note: This uses internally-tagged serde which typeshare doesn't support.
+/// The TypeScript type is manually defined in types-manual.ts.
 #[derive(Debug, Clone, Serialize, Deserialize, PartialEq)]
 #[serde(tag = "type", rename_all = "snake_case")]
 pub enum UnifiedAgentEvent {
